moved sources to sources.go from main.go
